Change Line classes inheritance tree structure
So that an Addition (or a Deletion) is not a kind of Context.

Before:

```
GitDiff::Line::Context
|- GitDiff::Line::Addition
`- GitDiff::Line::Deletion
```

After:

```
GitDiff::Line::Base
|- GitDiff::Line::Addition
|- GitDiff::Line::Context
`- GitDiff::Line::Deletion
```

The reason why I added this change is to check if a line is a context line or not.
