+++ lib/git_diff/line.rb
@@ -1,3 +1,4 @@
+require "git_diff/line/base"
 require "git_diff/line/context"
 require "git_diff/line/addition"
 require "git_diff/line/deletion"
