+++ lib/git_diff/line/addition.rb
@@ -1,6 +1,6 @@
 module GitDiff
   module Line
-    class Addition < Context
+    class Addition < Base
       def line_number=(line_number)
         @line_number = LineNumber.for_addition(line_number)
       end
