+++ lib/git_diff/line/deletion.rb
@@ -1,6 +1,6 @@
 module GitDiff
   module Line
-    class Deletion < Context
+    class Deletion < Base
       def line_number=(line_number)
         @line_number = LineNumber.for_deletion(line_number)
       end
